rename lib to lib_prompt_fusion
