+++ lib_prompt_fusion/hijacker.py
