+++ scripts/promptlang.py
@@ -3,9 +3,9 @@ import sys
 base_dir = scripts.basedir()
 sys.path.append(base_dir)
 
-from lib.interpolation_tensor import InterpolationTensorBuilder
-from lib.dsl_prompt_transpiler import parse_prompt
-from lib.hijacker import ModuleHijacker
+from lib_prompt_fusion.interpolation_tensor import InterpolationTensorBuilder
+from lib_prompt_fusion.dsl_prompt_transpiler import parse_prompt
+from lib_prompt_fusion.hijacker import ModuleHijacker
 import modules
 from modules.script_callbacks import on_script_unloaded
 from modules.prompt_parser import ScheduledPromptConditioning
